check whether file in use before delete

+++ pygmtools/dataset.py
@@ -30,6 +30,7 @@ import json
 import scipy.io as sio
 import glob
 import random
+import psutil
 from pygmtools.utils import download
 
 
@@ -224,7 +225,8 @@ class PascalVOC:
             for file_name in tqdm(file_names):
                 tar.extract(file_name, "data/PascalVOC/")
             tar.close()
-            os.remove(filename)
+            if not is_file_in_use(filename):
+                os.remove(filename)
 
         if name == "PascalVOC":
             print('Downloading dataset PascalVOC...')
@@ -243,7 +245,8 @@ class PascalVOC:
             for file_name in tqdm(file_names):
                 tar.extract(file_name, "data/PascalVOC/")
             tar.close()
-            os.remove(filename)
+            if not is_file_in_use(filename):
+                os.remove(filename)
         return filename
 
     def __filter_list(self, a_xml_list):
@@ -524,9 +527,7 @@ class WillowObject:
             os.makedirs(dirs)
 
         print('Downloading dataset WillowObject...')
-        import random
-        rand_number = random.randint(0,10000)
-        filename = "data/WILLOW_{}.zip".format(rand_number)
+        filename = "data/WILLOW.zip"
         download(filename=filename, url=url, to_cache=False)
         try:
             fz = zipfile.ZipFile(filename, "r")
@@ -539,7 +540,8 @@ class WillowObject:
         sleep(0.5)
         for file in tqdm(fz.namelist()):
             fz.extract(file, "data/WillowObject/")
-        os.remove(filename)
+        if not is_file_in_use(filename):
+            os.remove(filename)
         return filename
 
     def process(self):
@@ -818,7 +820,8 @@ class SPair71k:
         for file_name in tqdm(file_names):
             tar.extract(file_name, "data/")
         tar.close()
-        os.remove(filename)
+        if not is_file_in_use(filename):
+            os.remove(filename)
         return filename
 
     def process(self):
@@ -1067,7 +1070,8 @@ class IMC_PT_SparseGM:
         for file_name in tqdm(file_names):
             tar.extract(file_name, "data/")
         tar.close()
-        os.remove(filename)
+        if not is_file_in_use(filename):
+            os.remove(filename)
         return filename
 
     def process(self):
@@ -1260,7 +1264,8 @@ class CUB2011:
         for file_name in tqdm(file_names):
             tar.extract(file_name, "data/")
         tar.close()
-        os.remove(filename)
+        if not is_file_in_use(filename):
+            os.remove(filename)
         return filename
 
     def process(self):
@@ -1344,3 +1349,19 @@ class CUB2011:
             else:
                 ans[keys[idx]] = [val_i]
         return ans
+
+
+def is_file_in_use(file_path):
+    """
+    Check if a file is currently in use by any process
+    """
+    abs_path = os.path.abspath(file_path)
+
+    for proc in psutil.process_iter():
+        try:
+            for item in proc.open_files():
+                if abs_path == item.path:
+                    return True
+        except (psutil.NoSuchProcess, psutil.AccessDenied, psutil.ZombieProcess):
+            pass
+    return False
